Switched to use svn export command.

+++ release.py
@@ -30,8 +30,6 @@ import glob
 import os
 import os.path
 import shutil
-from subprocess import Popen
-from subprocess import PIPE
 from subprocess import call
 
 # The root of the source archive corresponds to the directory in which this
@@ -47,41 +45,6 @@ import version
 REL_NAME = "%s-%s" % (about.APPLICATION_NAME.lower(), version.get_version())
 REL_NAME_ZIP = "%s.zip" % REL_NAME
 
-def ignore(file):
-    if os.path.basename(file).startswith("."):
-        return True
-    if os.path.basename(file).endswith(".pyc"):
-        return True
-    if os.path.basename(file).endswith("~"):
-        return True
-    if os.path.basename(file).endswith(".zip"):
-        return True
-    if os.path.basename(file).endswith("tags"):
-        return True
-    return False
-
-def generate_filelist(subpath=""):
-    current_path = os.path.join(ROOT_DIR, subpath) or "."
-    if not os.path.isdir(current_path):
-        return [subpath]
-    filelist = []
-    for file in os.listdir(current_path):
-        file_subpath = os.path.join(subpath, file)
-        if not ignore(file_subpath):
-            filelist.extend(generate_filelist(file_subpath))
-        else:
-            print("Ignoring %s" % file_subpath)
-    return filelist
-
-def copy_files(files):
-    for path in files:
-        dest = os.path.join(REL_NAME, os.path.dirname(path))
-        if dest and not os.path.exists(dest):
-            print("Creating directory %s" % dest)
-            os.makedirs(dest)
-        print("Copying %s" % path)
-        shutil.copy(os.path.join(ROOT_DIR, path), dest)
-
 if os.path.exists(REL_NAME):
     print("Error: Path already exists %s" % REL_NAME)
     sys.exit()
@@ -90,18 +53,29 @@ if os.path.exists(REL_NAME_ZIP):
     print("Error: Package already exists %s" % REL_NAME_ZIP)
     sys.exit()
 
+print("Exporting from SVN")
+if call(["svn", "export", ROOT_DIR, REL_NAME]) != 0:
+    print("Error: Could not export from SVN")
+    sys.exit()
+
 print("Making manual")
 if call(["make", "-C", os.path.join(ROOT_DIR, "manual")]) != 0:
     print("Error: Could not make manual")
     sys.exit()
 
-print("Creating directory %s" % REL_NAME)
-os.mkdir(REL_NAME)
-
-copy_files(generate_filelist())
+# The html version of the manual is not versioned controlled, but we would like
+# to include it in the source release anyway so that users running from source
+# do not need to make it themselves.
+print("Copying manual to release")
+if call(["cp", os.path.join(ROOT_DIR, "manual", "manual.html"),
+         os.path.join(REL_NAME, "manual")]) != 0:
+    print("Error: Could not copy manual")
+    sys.exit()
 
 print("Creating file %s" % REL_NAME_ZIP)
-Popen(["zip", "-r", REL_NAME_ZIP, REL_NAME], stdout=PIPE).communicate()[0]
+if call(["zip", "-r", REL_NAME_ZIP, REL_NAME]) != 0:
+    print("Error: Could not create archive")
+    sys.exit()
 
 print("Deleting directory %s" % REL_NAME)
 shutil.rmtree(REL_NAME)
